Refactor into separate component files for clarity
